refactor: Refactored startup using minimal hosting model with extension methods
