+++ src/Eurofurence.App.Server.Services/Abstractions/JobOption.cs
@@ -1,0 +1,9 @@
+﻿namespace Eurofurence.App.Server.Services.Abstractions
+{
+    public class JobOption
+    {
+        public bool Enabled { get; set; }
+
+        public int SecondsInterval { get; set; }
+    }
+}
