+++ src/Eurofurence.App.Server.Services/Abstractions/Telegram/TelegramOptions.cs
@@ -1,6 +1,6 @@
 namespace Eurofurence.App.Server.Services.Abstractions.Telegram;
 
-public class TelegramConfiguration
+public class TelegramOptions
 {
     public string AccessToken { get; set; }
 }
